requirements.txt
fastapi==0.115.8
passlib==1.7.4
python-jose==3.3.0
uvicorn==0.34.0
SQLAlchemy==2.0.38
aiosqlite==0.21.0
pydantic[email]==2.10.6
alembic==1.14.1
python-dotenv==1.0.1
